feat: Add Blogify - A Blogging Platform to resume.pdf
Add a new section to the resume.pdf file showcasing the Blogify project. This project is a blogging platform that allows users to write articles for different blogs and easily share their thoughts and stories. The section includes a brief description of the project and a link to the website.